+++ backend/src/main/java/com/sdacademy/budgettracker/repository/IncomeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IncomeRepository extends JpaRepository<Income, Long> {
+public interface IncomeRepository extends JpaRepository<Income, Double> {
     @Query("SELECT SUM(i.amount) FROM Income i WHERE i.user.id = :userId")
-    Optional<Double> findTotalIncomeByUserId(@Param("userId") Double userId);
+    Optional<Double> findTotalIncomeByUserId(@Param("userId") Long userId);
 }
