+++ backend/src/main/java/com/sdacademy/budgettracker/service/BudgetTrackerService.java
@@ -7,12 +7,12 @@ public interface BudgetTrackerService {
     void createRecord(BudgetTrackerRecordDTO recordDTO);
     void login(String username, String password);
     void registerUser(User user);
-    void enterInitialIncome(Double totalIncome, Double userId);
-    void setSavingsGoal(double goalPercentage, Double userId);
-    Double getTotalIncome(Double userId);
-    Double calculateTotalExpenses(Double userId);
+    void enterInitialIncome(Double totalIncome, Long userId);
+    void setSavingsGoal(double goalPercentage, Long userId);
+    Double getTotalIncome(Long userId);
+    Double calculateTotalExpenses(Long userId);
     void addExpense(BudgetTrackerRecordDTO recordDTO);
     void addAdditionalIncome(BudgetTrackerRecordDTO recordDTO);
-    void checkSavingsStatus(Double userId);
-    Object getOverviewOfExpenses(Double userId);
+    void checkSavingsStatus(Long userId);
+    Object getOverviewOfExpenses(Long userId);
 }
